Revert "SubstratumService: Refactor & cleanup according to AOSP conventions"
This reverts commit 78f5ec499449774c257c17f8fd71cc207d9becbe.
