+++ packages/SubstratumService/AndroidManifest.xml
@@ -19,26 +19,26 @@
 
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="projekt.substratum.helper"
+    package="android.substratum.service"
     android:sharedUserId="android.uid.system"
     android:versionCode="2"
     android:versionName="two" >
 
-    <uses-permission android:name="android.permission.INTERACT_ACROSS_USERS_FULL"/>
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
-    
+    <uses-permission android:name="android.permission.INTERACT_ACROSS_USERS_FULL"/>
+    <uses-permission android:name="android.permission.CHANGE_OVERLAY_PACKAGES"/>
     <uses-feature name="projekt.substratum.theme" required="true"/>
 
     <application
         android:allowBackup="false"
-        android:label="Substratum Helper Service">
+        android:label="Substratum Service">
 
-        <service android:name="projekt.substratum.helper.SubstratumHelperService"
+        <service android:name="android.substratum.service.SubstratumHelperService"
                  android:enabled="true"
                  android:exported="true">
             <intent-filter>
-                <action android:name="projekt.substratum.helper.SubstratumHelperService" />
+                <action android:name="android.substratum.service.SubstratumHelperService" />
             </intent-filter>
         </service>
